pyproject.toml
[tool.poetry]
name = "fragalysis-backend"
version = "2.0.0"
description = "The backend for the Fragalysis Stack web application."
authors = ["Rachael Skyner <rachael.skyner@diamond.ac.uk>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "fragalysis_backend"}]

[tool.poetry.dependencies]
python = "^3.11.3"

# Specific (critical dependency) packages...
# Do not change these without careful thought!
rdkit = "2023.3.2"
xchem-db = "0.1.26b0"

# Less strict (flexible dependency) packages...
celery = "^5.3.1"
deepdiff = "^6.2.0"
djangorestframework = "^3.14.0"
django-bootstrap3 = "^23.4"
django-cleanup = "^8.0.0"
django-extensions = "^3.2.3"
django-filter = "^23.2"
django-guardian = "^2.4.0"
django-simple-history = "^3.3.0"
django-webpack-loader = "^0.7.0"
django-rest-swagger = "^2.2.0"
fragalysis = "^1.1.0"
graphene-django = "^3.1.3"
gunicorn = "^21.2.0"
im-squonk2-client = "^1.22.2"
mozilla_django_oidc = "^3.0.0"
mysql-connector-python = "^8.1.0"
pandas = "^1.5.3"
pandoc = "^2.3"
psutil = "^5.9.5"
psycopg2-binary = "^2.9.6"
pydiscourse = "^1.3.0"
pymysql = "^1.1.0"
redis = "^4.6.0"
sentry-sdk = "^1.29.2"
shortuuid = "^1.0.11"
sshtunnel = "^0.4.0"
urllib3 = "^2.0.4"
validators = "^0.20.0"

# Blocked packages...
#
bcrypt = "4.0.1"
# Django will (probably) need work in the code to support v4.
# Simply changing to django 4 results in errors
# like "CSRF Failed: Origin checking failed"
django = "3.2.20"
# We're stuck on ISPyB 4.x to avoid the error: -
#   ImportError: cannot import name 'ISPyBNoResultException' from 'ispyb.connector.mysqlsp.main'
ispyb = "4.13.2"
# ispyb (4.13.2) depends on mysql-connector (<2.2.3)
mysql-connector = "2.1.7"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pylint = "^2.17.4"
pylint-django = "^2.5.3"
httpie = "^3.2.2"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
